% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dsem_sensitvity.R
\name{run_dsem_sensitivity}
\alias{run_dsem_sensitivity}
\title{Run DSEM Sensitivity Analysis Workflow}
\usage{
run_dsem_sensitivity(
  weights_file,
  adj_file,
  lags_file,
  node_classes,
  latent_dict = NULL,
  out_dir,
  base_mu,
  base_sd,
  base_slope,
  max_ts_length = 400,
  M_preflight = 30,
  N_rand = 30,
  cores = 1,
  missing_fractions = c(0, 0.1, 0.25, 0.5),
  variance_multipliers = c(1, 2, 5, 10),
  ts_length_fractions = c(1, 0.5, 0.25, 0.1),
  detrend = detrend,
  standardize = standardize,
  control_time_drift = control_time_drift,
  jumpstart = FALSE
)
}
\description{
Run DSEM Sensitivity Analysis Workflow
}
